Add support for NHL 07 on PSP, add room handling, externalize game-specific logic
