+++ src/main/java/com/ea/dto/SocketWrapper.java
@@ -7,6 +7,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import lombok.extern.slf4j.Slf4j;
 
 import java.net.Socket;
 import java.util.concurrent.atomic.AtomicBoolean;
@@ -15,14 +16,30 @@ import java.util.concurrent.atomic.AtomicBoolean;
 @NoArgsConstructor
 @Getter
 @Setter
+@Slf4j
 public class SocketWrapper {
+    private final AtomicBoolean isDedicatedHost = new AtomicBoolean(false);
+    private final AtomicBoolean isGps = new AtomicBoolean(false);
+    private final AtomicBoolean isHosting = new AtomicBoolean(false);
     private Socket socket;
     private String identifier;
     private volatile String lkey;
-    private final AtomicBoolean isHost = new AtomicBoolean(false);
-    private final AtomicBoolean isGps = new AtomicBoolean(false);
-    private final AtomicBoolean isHosting = new AtomicBoolean(false);
+    private volatile String userflags = "1";
     private volatile AccountEntity accountEntity;
     private volatile PersonaEntity personaEntity;
     private volatile PersonaConnectionEntity personaConnectionEntity;
+    private volatile Thread gameSearchThread;
+
+
+    public void cleanupOnSocketClose(SocketWrapper socketWrapper) {
+        synchronized (this) {
+            Thread searchThread = socketWrapper.getGameSearchThread();
+            if (searchThread != null && searchThread.isAlive()) {
+                log.debug("Game search thread for socket {} finished", socket.getRemoteSocketAddress());
+                searchThread.interrupt();
+                socketWrapper.setGameSearchThread(null);
+            }
+        }
+    }
+
 }
