+++ src/main/java/com/ea/services/core/GameService.java
@@ -1,15 +1,19 @@
 package com.ea.services.core;
 
+import com.ea.dto.Room;
 import com.ea.dto.SocketData;
 import com.ea.dto.SocketWrapper;
-import com.ea.entities.core.*;
-import com.ea.entities.stats.MohhPersonaStatsEntity;
+import com.ea.entities.core.AccountEntity;
+import com.ea.entities.core.GameConnectionEntity;
+import com.ea.entities.core.GameEntity;
+import com.ea.entities.core.PersonaConnectionEntity;
 import com.ea.mappers.SocketMapper;
 import com.ea.repositories.core.*;
-import com.ea.repositories.stats.MohhPersonaStatsRepository;
 import com.ea.services.server.GameServerService;
 import com.ea.services.server.SocketManager;
+import com.ea.services.stats.MohhStatsService;
 import com.ea.steps.SocketWriter;
+import com.ea.utils.GameUtils;
 import com.ea.utils.SocketUtils;
 import jakarta.annotation.PostConstruct;
 import jakarta.annotation.PreDestroy;
@@ -20,14 +24,11 @@ import org.springframework.stereotype.Service;
 
 import java.net.Socket;
 import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static com.ea.services.server.GameServerService.*;
-import static com.ea.utils.HexUtils.*;
-import static com.ea.utils.SocketUtils.DATETIME_FORMAT;
 import static com.ea.utils.SocketUtils.getValueFromSocket;
 
 @Slf4j
@@ -38,44 +39,202 @@ public class GameService {
     private final GameRepository gameRepository;
     private final GameConnectionRepository gameConnectionRepository;
     private final PersonaConnectionRepository personaConnectionRepository;
-    private final MohhPersonaStatsRepository mohhPersonaStatsRepository;
     private final AccountRepository accountRepository;
     private final BlacklistRepository blacklistRepository;
     private final SocketMapper socketMapper;
     private final PersonaService personaService;
+    private final GameServerService gameServerService;
+    private final RoomService roomService;
+    private final MohhStatsService mohhStatsService;
     private final SocketWriter socketWriter;
     private final SocketManager socketManager;
-    private final GameServerService gameServerService;
+    private final GameUtils gameUtils;
+
 
     /**
-     * Distribute room change updates
+     * Join the best matching game based on provided criteria
+     * Most arguments are game-specific, but the MODE defines how the search will be performed
+     * Modes are :
+     * - LOBBYAPI_GQWK_MODE_FAIL   - Fail with EC_NOT_FOUND error code
+     * - LOBBYAPI_GQWK_MODE_CREATE - Create a game
+     * - LOBBYAPI_GQWK_MODE_WAIT   - Wait until a game or user becomes
+     * - LOBBYAPI_GQWK_MODE_CANCEL - Cancel a MODE_WAIT request
+     * <p>
+     * NHL uses MODE=2, so it will wait until a game is available,
+     * and it will send MODE=3 when the user wants to cancel the request
      *
-     * @param socket     The socket to write the response to
-     * @param socketData The socket data
+     * @param socket        The socket to write the response to
+     * @param socketData    The socket data containing game search criteria
+     * @param socketWrapper The socket wrapper of current connection
      */
-    public void rom(Socket socket, SocketData socketData) {
-        Map<String, String> content = Stream.of(new String[][]{
-                {"I", "1"}, // Room identifier
-                {"N", "room"}, // Room name
-//                { "H", socketManager.getSocketWrapper(socket.getRemoteSocketAddress().toString()).getPers() }, // Room Host
-//                { "D", "" }, // Room description
-//                { "F", "CK" }, // Attribute flags
-//                { "T", "1" }, // Current room population
-//                { "L", "33" }, // Max users allowed in room
-//                { "P", "0" }, // Room ping
-//                { "A", props.getTcpHost() }, // Room address
-        }).collect(Collectors.toMap(data -> data[0], data -> data[1]));
-
-        socketData.setOutputData(content);
-        socketData.setIdMessage("+rom");
+    public void gqwk(Socket socket, SocketData socketData, SocketWrapper socketWrapper) {
         socketWriter.write(socket, socketData);
+        String mode = getValueFromSocket(socketData.getInputMessage(), "MODE");
+
+        if ("2".equals(mode)) { // Wait for a game to become available
+            // Cancel any existing search thread for this socket
+            synchronized (socketWrapper) {
+                Thread existingThread = socketWrapper.getGameSearchThread();
+                if (existingThread != null && existingThread.isAlive()) {
+                    existingThread.interrupt();
+                }
+            }
+
+            // Create a new thread to handle the game search with timer
+            Thread searchThread = new Thread(() -> {
+                String vers = socketWrapper.getPersonaConnectionEntity().getVers();
+                List<String> relatedVers = gameServerService.getRelatedVers(vers);
+
+                while (!Thread.currentThread().isInterrupted() && !socket.isClosed()) {
+                    log.debug("Searching for game for socket: {}", socket.getRemoteSocketAddress());
+                    List<GameEntity> gameEntities = gameRepository.findByVersInAndEndTimeIsNull(relatedVers);
+
+                    // Game must not be full, not use pw, not be started, and have at least one active game connection
+                    gameEntities = gameEntities.stream()
+                            .filter(gameEntity -> !gameEntity.isStarted() && gameEntity.getGameConnections().stream()
+                                    .anyMatch(connection -> connection.getEndTime() == null))
+                            .filter(gameEntity -> gameEntity.getGameConnections().stream()
+                                    .filter(connection -> connection.getEndTime() == null).count() < gameEntity.getMaxsize())
+                            .filter(gameEntity -> StringUtils.isEmpty(gameEntity.getPass()))
+                            .toList();
+
+                    // Theoretically we should filter the game entities based on the criteria provided in the socket data
+                    // Given the low player count, no need to filter by params
+
+                    if (!gameEntities.isEmpty()) {
+                        if (!socket.isClosed()) {
+                            GameEntity gameEntity = gameEntities.get(0); // Get the first game found
+                            // It seems like the "Play Now" features is auto-start, it doesn't join the lobby with only +mgm
+                            joinGame(socket, socketData, socketWrapper, gameEntity);
+                            gsta(socket, socketData, socketWrapper);
+                        }
+                        break; // End the thread when a game is found
+                    }
+
+                    try {
+                        Thread.sleep(15000); // Wait 15 seconds before next search
+                    } catch (InterruptedException e) {
+                        Thread.currentThread().interrupt();
+                        break;
+                    }
+                }
+
+                // Clean up the thread reference when done
+                synchronized (socketWrapper) {
+                    if (socketWrapper.getGameSearchThread() == Thread.currentThread()) {
+                        log.debug("Game search thread for socket {} finished", socket.getRemoteSocketAddress());
+                        socketWrapper.setGameSearchThread(null);
+                    }
+                }
+            });
+
+            // Store the search thread for this socket wrapper
+            synchronized (socketWrapper) {
+                socketWrapper.setGameSearchThread(searchThread);
+            }
+            searchThread.start();
+
+        } else if ("3".equals(mode)) {
+            // Cancel a MODE_WAIT request
+            synchronized (socketWrapper) {
+                Thread searchThread = socketWrapper.getGameSearchThread();
+                if (searchThread != null && searchThread.isAlive()) {
+                    log.debug("Cancelling game search for socket: {}", socket.getRemoteSocketAddress());
+                    searchThread.interrupt();
+                    socketWrapper.setGameSearchThread(null);
+                }
+            }
+        }
     }
 
-    public void gsta(Socket socket, SocketData socketData) {
+    /**
+     * Start the game
+     *
+     * @param socket        The socket to write the response to
+     * @param socketData    The socket data
+     * @param socketWrapper The socket wrapper of current connection
+     */
+    public void gsta(Socket socket, SocketData socketData, SocketWrapper socketWrapper) {
         socketWriter.write(socket, socketData);
+
+        // If game is P2P then we send +ses to all players in the lobby
+        if (gameServerService.isP2P(socketWrapper.getPersonaConnectionEntity().getVers())) {
+            GameConnectionEntity gameConnectionEntity = gameConnectionRepository.findByPersonaConnectionIdAndEndTimeIsNull(
+                    socketWrapper.getPersonaConnectionEntity().getId()).orElse(null);
+
+            if (gameConnectionEntity != null) {
+                GameEntity gameEntity = gameConnectionEntity.getGame();
+                if (gameEntity != null) {
+                    for (GameConnectionEntity connection : gameConnectionRepository.findByGameIdAndEndTimeIsNull(gameEntity.getId())) {
+                        SocketWrapper connectionSocketWrapper = socketManager.getSocketWrapperByPersonaConnectionId(connection.getPersonaConnection().getId());
+                        if (connectionSocketWrapper != null) {
+                            ses(connectionSocketWrapper.getSocket(), gameEntity);
+                        }
+                        connection.setStartTime(LocalDateTime.now());
+                        gameConnectionRepository.save(connection);
+                    }
+                    // Don't update start time here, the WHEN attribute of the 'rank' packet uses the first declared start time (and it's used as an identifier for the game)
+                    gameEntity.setStarted(true);
+                    gameRepository.save(gameEntity);
+                }
+            }
+        }
     }
 
+    /**
+     * Set game parameters
+     * This is used to update the game parameters, such as name, params, sysflags, etc.
+     * It is also used to on map rotation on dedicated servers, so it will end the current game and create a new one with the new parameters.
+     *
+     * @param socket        The socket to write the response to
+     * @param socketData    The socket data
+     * @param socketWrapper The socket wrapper of current connection
+     */
     public void gset(Socket socket, SocketData socketData, SocketWrapper socketWrapper) {
+        socketWriter.write(socket, socketData);
+
+        PersonaConnectionEntity personaConnectionEntity = socketWrapper.getPersonaConnectionEntity();
+        if (gameServerService.isP2P(personaConnectionEntity.getVers())) {
+            // On NHL, the value is 0 or 1 to know if the client is ready or not
+            String userflags = getValueFromSocket(socketData.getInputMessage(), "USERFLAGS");
+
+            if (userflags != null) {
+                synchronized (this) {
+                    socketWrapper.setUserflags(userflags);
+                }
+                // Broadcast the userflags to all players in the game
+                Optional<GameConnectionEntity> gameConnectionOpt = gameConnectionRepository.findByPersonaConnectionIdAndEndTimeIsNull(personaConnectionEntity.getId());
+
+                if (gameConnectionOpt.isPresent()) {
+                    GameConnectionEntity gameConnectionEntity = gameConnectionOpt.get();
+                    GameEntity gameEntity = gameConnectionEntity.getGame();
+                    List<GameConnectionEntity> gameConnections = gameConnectionRepository.findByGameIdAndEndTimeIsNull(gameEntity.getId());
+                    for (GameConnectionEntity firstLevelConnection : gameConnections) { // For each player in the game, send +agm and +mgm
+                        SocketWrapper firstLevelSocketWrapper = socketManager.getSocketWrapperByPersonaConnectionId(firstLevelConnection.getPersonaConnection().getId());
+                        if (firstLevelSocketWrapper != null) {
+                            agm(firstLevelSocketWrapper.getSocket(), gameEntity);
+                            mgm(firstLevelSocketWrapper.getSocket(), gameEntity);
+                        }
+                    }
+                }
+            }
+        }
+
+        // For MoHH dedicated server, gset means a map rotation, but instead of just updating the game parameters,
+        // we end the current game and create a new one with the new parameters.
+        if (MOH07_OR_UHS.contains(personaConnectionEntity.getVers())) {
+            handleMapRotation(socketData, socketWrapper);
+        }
+    }
+
+    /**
+     * Handle map rotation for MoHH dedicated servers
+     * This will end the current game and create a new one with the new parameters.
+     *
+     * @param socketData    The socket data
+     * @param socketWrapper The socket wrapper of current connection
+     */
+    private void handleMapRotation(SocketData socketData, SocketWrapper socketWrapper) {
         //String name = getValueFromSocket(socketData.getInputMessage(), "NAME");
         String params = getValueFromSocket(socketData.getInputMessage(), "PARAMS");
         String sysflags = getValueFromSocket(socketData.getInputMessage(), "SYSFLAGS");
@@ -84,8 +243,6 @@ public class GameService {
                         socketWrapper.getPersonaConnectionEntity().getId())
                 .filter(GameConnectionEntity::isHost)
                 .map(GameConnectionEntity::getGame).orElse(null);
-
-        socketWriter.write(socket, socketData);
 
         LocalDateTime now = LocalDateTime.now();
         new Thread(() -> {
@@ -111,6 +268,8 @@ public class GameService {
                     newGameEntity.setPass(gameEntity.getPass());
                     newGameEntity.setMinsize(gameEntity.getMinsize());
                     newGameEntity.setMaxsize(gameEntity.getMaxsize());
+                    newGameEntity.setStarted(true);
+                    newGameEntity.setRoomId(gameEntity.getRoomId());
                     gameRepository.save(newGameEntity);
 
                     for (GameConnectionEntity gameConnectionEntity : gameConnections) {
@@ -151,6 +310,15 @@ public class GameService {
         gam(socket, filteredGameEntities);
     }
 
+    /**
+     * Filter game entities based on the criteria provided in the paramsMap
+     * Each game version may have different criteria for filtering games
+     *
+     * @param gameEntities The list of game entities to filter
+     * @param paramsMap    The parameters map from the socket data
+     * @param vers         The version of the game
+     * @return A list of filtered game entities
+     */
     private List<GameEntity> filterGameEntities(List<GameEntity> gameEntities, Map<String, String> paramsMap, String vers) {
         int count = Integer.parseInt(paramsMap.get("COUNT"));
         return gameEntities.stream()
@@ -159,105 +327,21 @@ public class GameService {
                 .toList();
     }
 
+    /**
+     * Check if the game entity matches the criteria based on the version
+     * Each game version may have different criteria for filtering games
+     *
+     * @param gameEntity The game entity to check
+     * @param paramsMap  The parameters map from the socket data
+     * @param vers       The version of the game
+     * @return true if the game entity matches the criteria, false otherwise
+     */
     private boolean matchesCriteria(GameEntity gameEntity, Map<String, String> paramsMap, String vers) {
-        String name = paramsMap.get("NAME");
-        String available = paramsMap.getOrDefault("AVAILABLE", paramsMap.get("AVAIL")); // Game not full (AVAIL on MoHH2)
-        String mode = paramsMap.get("MODE");
-        String map = paramsMap.get("MAP");
-        String ff = paramsMap.get("FF"); // Friendly fire
-        String aim = paramsMap.get("AIM"); // Aim assist
-        String ctrl = paramsMap.get("CTRL"); // Elite or Wii Zapper
-        String maxsize = paramsMap.get("MAXSIZE");
-        String sysmaskParam = paramsMap.get("SYSMASK");
-        String sysflagsParam = paramsMap.get("SYSFLAGS");
-        String tb = paramsMap.get("TB"); // Team balance
-        String ak = paramsMap.get("AK"); // Auto-kick
-        String smg = paramsMap.get("SMG");
-        String hmg = paramsMap.get("HMG");
-        String rif = paramsMap.get("RIF");
-        String snip = paramsMap.get("SNIP");
-        String shotg = paramsMap.get("SHOTG");
-        String baz = paramsMap.get("BAZ");
-        String gren = paramsMap.get("GREN");
-
-        String nameDb = gameEntity.getName().replaceAll("\"", "").toLowerCase();
-        boolean availableDb = gameEntity.getGameConnections().stream().filter(connection -> null == connection.getEndTime()).count() < gameEntity.getMaxsize();
-        String[] params = gameEntity.getParams().split(",");
-        String modeDb = params[0];
-        int mapDb = Integer.parseInt(params[1], 16); // The map in db is in hex, so we need to convert it to decimal
-        boolean hasPwDb = gameEntity.getPass() != null;
-        String ffDb = params[2];
-
-        String aimDb = null, ctrlDb = null, tbDb = null, akDb = null;
-        switch (vers) {
-            case PSP_MOH_07 -> aimDb = params[3];
-            case PSP_MOH_08 -> {
-                aimDb = params[3];
-                tbDb = params[4];
-                akDb = params[9];
-            }
-            case WII_MOH_08 -> {
-                tbDb = params[3];
-                akDb = params[8];
-                ctrlDb = params[9];
-            }
-        }
-
-        boolean isRankedDb = vers.equals(PSP_MOH_07) ? StringUtils.isNotEmpty(params[8]) : StringUtils.isNotEmpty(params[17]);
-
-        String smgDb = null, hmgDb = null, rifDb = null, snipDb = null, shotgDb = null, bazDb = null, grenDb = null;
-        if (vers.equals(PSP_MOH_08) || vers.equals(WII_MOH_08)) {
-            smgDb = params[10];
-            hmgDb = params[11];
-            rifDb = params[12];
-            snipDb = params[13];
-            shotgDb = params[14];
-            bazDb = params[15];
-            grenDb = params[16];
-        }
-
-        boolean matchesFlags = matchesFlags(sysmaskParam, sysflagsParam, hasPwDb, isRankedDb);
-
-        return (name == null || nameDb.contains(name.replaceAll("\"", "").toLowerCase()))
-                && (available.equals("-1") || availableDb)
-                && (mode.equals("-1") || modeDb.equals(mode))
-                && (map.equals("-1") || mapDb == Integer.parseInt(map))
-                && (ff.equals("-1") || (ff.equals("0") && ffDb.equals("")) || ffDb.equals(ff))
-                && (aim == null || aim.equals("-1") || (aim.equals("0") && aimDb.equals("")) || aimDb.equals(aim))
-                && (ctrl == null || ctrl.equals("-1") || (ctrl.equals("0") && ctrlDb.equals("")) || ctrlDb.equals(ctrl))
-                && (maxsize == null || gameEntity.getMaxsize() == Integer.parseInt(maxsize))
-                && matchesFlags
-                && (tb == null || tb.equals("-1") || (tb.equals("0") && tbDb.equals("")) || tbDb.equals(tb))
-                && (ak == null || ak.equals("-1") || (ak.equals("0") && akDb.equals("")) || akDb.equals(ak))
-                && (smg == null || smg.equals("-1") || (smg.equals("0") && smgDb.equals("")) || smgDb.equals(smg))
-                && (hmg == null || hmg.equals("-1") || (hmg.equals("0") && hmgDb.equals("")) || hmgDb.equals(hmg))
-                && (rif == null || rif.equals("-1") || (rif.equals("0") && rifDb.equals("")) || rifDb.equals(rif))
-                && (snip == null || snip.equals("-1") || (snip.equals("0") && snipDb.equals("")) || snipDb.equals(snip))
-                && (shotg == null || shotg.equals("-1") || (shotg.equals("0") && shotgDb.equals("")) || shotgDb.equals(shotg))
-                && (baz == null || baz.equals("-1") || (baz.equals("0") && bazDb.equals("")) || bazDb.equals(baz))
-                && (gren == null || gren.equals("-1") || (gren.equals("0") && grenDb.equals("")) || grenDb.equals(gren));
-    }
-
-
-    private boolean matchesFlags(String sysmaskParam, String sysflagsParam, boolean hasPwDb, boolean isRankedDb) {
-        if (sysmaskParam == null) {
+        if (vers.equals(PSP_MOH_07_UHS) || MOH07_OR_MOH08.contains(vers)) {
+            return mohhStatsService.matchesCriteria(gameEntity, paramsMap, vers);
+        } else {
             return true;
         }
-
-        int clientMask = Integer.parseInt(sysmaskParam);
-        int clientFlags = sysflagsParam != null ? Integer.parseInt(sysflagsParam) : 0;
-
-        // Calculate game's flags based on ranked and password status
-        int gameFlags = 0;
-        if (hasPwDb) {
-            gameFlags |= (1 << 16);  // Set bit 16 for password
-        }
-        if (isRankedDb) {
-            gameFlags |= (1 << 18);  // Set bit 18 for ranked
-        }
-
-        // Only check the bits that the client cares about (specified in mask)
-        return (gameFlags & clientMask) == (clientFlags & clientMask);
     }
 
     /**
@@ -307,17 +391,69 @@ public class GameService {
         }
 
         String ident = getValueFromSocket(socketData.getInputMessage(), "IDENT");
-        Optional<GameEntity> gameEntityOpt = gameRepository.findById(Long.valueOf(ident));
+        Optional<GameEntity> gameEntityOpt;
+        if (ident != null) {
+            gameEntityOpt = gameRepository.findById(Long.valueOf(ident));
+        } else {
+            // Some games don't provide an identifier, so we need to find the game by name and version
+            String name = getValueFromSocket(socketData.getInputMessage(), "NAME");
+            gameEntityOpt = gameRepository.findByNameAndVersInAndEndTimeIsNull(name, List.of(socketWrapper.getPersonaConnectionEntity().getVers()));
+        }
+
         if (gameEntityOpt.isPresent()) {
-            GameEntity gameEntity = gameEntityOpt.get();
-            String pass = getValueFromSocket(socketData.getInputMessage(), "PASS");
-            if (StringUtils.isNotEmpty(pass) && !pass.equals(gameEntity.getPass())) {
-                socketWriter.write(socket, new SocketData("gjoipass", null, null)); // Wrong password
+            joinGame(socket, socketData, socketWrapper, gameEntityOpt.get());
+        } else {
+            socketWriter.write(socket, new SocketData("gjoiugam", null, null)); // Game unknown
+        }
+    }
+
+    public void joinGame(Socket socket, SocketData socketData, SocketWrapper socketWrapper, GameEntity gameEntity) {
+        String pass = getValueFromSocket(socketData.getInputMessage(), "PASS");
+        if (StringUtils.isNotEmpty(pass) && !pass.equals(gameEntity.getPass())) {
+            socketWriter.write(socket, new SocketData("gjoipass", null, null)); // Wrong password
+            return;
+        }
+        if (gameEntity.getEndTime() == null) {
+            // Check if game allows joining mid-game
+            if (gameEntity.isStarted() && MIDGAME_FORBIDDEN.contains(gameEntity.getVers())) {
+                socketWriter.write(socket, new SocketData("gjoiasta", null, null)); // Game already started
                 return;
             }
-            if (gameEntity.getEndTime() == null) {
-                startGameConnection(socketWrapper, gameEntity);
-                socketWriter.write(socket, socketData);
+
+            // Check if the game is full
+            long currentCount = gameEntity.getGameConnections().stream().filter(connection -> null == connection.getEndTime()).count();
+            if (currentCount + 1 > gameEntity.getMaxsize()) {
+                socketWriter.write(socket, new SocketData("gjoifull", null, null)); // Game is full
+                return;
+            }
+
+            startGameConnection(socketWrapper, gameEntity, false);
+            socketWriter.write(socket, socketData);
+
+            // Check if game is P2P
+            if (gameServerService.isP2P(gameEntity.getVers())) {
+                // Reset userflags
+                synchronized (this) {
+                    socketWrapper.setUserflags("0");
+                }
+
+                // Send who
+                personaService.who(socket, socketWrapper);
+
+                // Broadcast the game join to all connected clients in the room
+                // Inform about all players in the game (+usr for each player, to each player)
+                List<GameConnectionEntity> gameConnections = gameConnectionRepository.findByGameIdAndEndTimeIsNull(gameEntity.getId());
+                for (SocketWrapper clientWrapper : socketManager.getSocketWrapperByVers(gameEntity.getVers())) {
+                    for (GameConnectionEntity gameConnection : gameConnections) { // For each player in the game, send +usr to each player
+                        SocketWrapper inGameWrapper = socketManager.getSocketWrapperByPersonaConnectionId(gameConnection.getPersonaConnection().getId());
+                        if (inGameWrapper != null && gameConnections.stream().anyMatch(gameConnectionEntity -> gameConnectionEntity.getPersonaConnection().getId().equals(clientWrapper.getPersonaConnectionEntity().getId()))) {
+                            personaService.usr(clientWrapper.getSocket(), inGameWrapper); // Update user info for each player
+                        }
+                    }
+                    agm(clientWrapper.getSocket(), gameEntity);
+                    mgm(clientWrapper.getSocket(), gameEntity);
+                }
+            } else {
                 updateHostInfo(gameEntity);
                 try {
                     Thread.sleep(100);
@@ -325,16 +461,14 @@ public class GameService {
                     Thread.currentThread().interrupt();
                 }
                 ses(socket, gameEntity);
-            } else {
-                socketWriter.write(socket, new SocketData("gjoiugam", null, null)); // Game closed
             }
         } else {
-            socketWriter.write(socket, new SocketData("gjoiugam", null, null)); // Game unknown
+            socketWriter.write(socket, new SocketData("gjoiugam", null, null)); // Game closed
         }
     }
 
     /**
-     * Create a game on a persistent game spawn service for a user
+     * Create a game on a persistent game spawn service for a user (dedicated servers)
      *
      * @param socket     The socket to write the response to
      * @param socketData The socket data
@@ -350,10 +484,11 @@ public class GameService {
 
         String vers = socketWrapper.getPersonaConnectionEntity().getVers();
         String slus = socketWrapper.getPersonaConnectionEntity().getSlus();
-        List<String> relatedVers = gameServerService.getRelatedVers(vers);
-        boolean isMohh = relatedVers.equals(VERS_MOHH_PSP);
+
         GameEntity gameEntityToCreate = socketMapper.toGameEntity(socketData.getInputMessage(), vers, slus);
 
+        List<String> relatedVers = gameServerService.getRelatedVers(vers);
+        boolean isMohh = relatedVers.equals(MOH07_OR_UHS);
         boolean duplicateName = gameRepository.existsByNameAndVersInAndEndTimeIsNull(gameEntityToCreate.getName(), relatedVers);
         if (duplicateName) {
             socketData.setIdMessage("gpscdupl");
@@ -386,7 +521,7 @@ public class GameService {
                         Optional<GameEntity> gameEntityOpt = gameRepository.findByNameAndVersInAndEndTimeIsNull(gameEntityToCreate.getName(), relatedVers);
                         if (gameEntityOpt.isPresent()) {
                             GameEntity gameEntity = gameEntityOpt.get();
-                            startGameConnection(socketWrapper, gameEntity);
+                            startGameConnection(socketWrapper, gameEntity, false);
                             ses(socket, gameEntity);
                             updateHostInfo(gameEntity);
                             break;
@@ -409,9 +544,10 @@ public class GameService {
                 }
             }
             gameEntityToCreate.setParams(params);
-
+            gameEntityToCreate.setStarted(true);
             gameRepository.save(gameEntityToCreate);
-            startGameConnection(socketWrapper, gameEntityToCreate);
+
+            startGameConnection(socketWrapper, gameEntityToCreate, false);
             ses(socket, gameEntityToCreate);
         }
     }
@@ -424,7 +560,7 @@ public class GameService {
     public void updateHostInfo(GameEntity gameEntity) {
         SocketWrapper hostSocketWrapper = socketManager.getHostSocketWrapperOfGame(gameEntity.getId());
         if (hostSocketWrapper != null) {
-            Map<String, String> content = getGameInfo(gameEntity);
+            Map<String, String> content = gameUtils.getGameInfo(gameEntity);
             socketWriter.write(hostSocketWrapper.getSocket(), new SocketData("+mgm", null, content));
             try {
                 Thread.sleep(100);
@@ -447,6 +583,11 @@ public class GameService {
         String slus = socketWrapper.getPersonaConnectionEntity().getSlus();
         GameEntity gameEntity = socketMapper.toGameEntity(socketData.getInputMessage(), vers, slus);
 
+        // Some games don't provide a game name, so we set it to the persona name
+        if (gameEntity.getName() == null || gameEntity.getName().isEmpty()) {
+            gameEntity.setName(socketWrapper.getPersonaEntity().getPers());
+        }
+
         List<String> relatedVers = gameServerService.getRelatedVers(vers);
         boolean duplicateName = gameRepository.existsByNameAndVersInAndEndTimeIsNull(gameEntity.getName(), relatedVers);
 
@@ -454,17 +595,40 @@ public class GameService {
             socketData.setIdMessage("gcredupl");
             socketWriter.write(socket, socketData);
         } else {
-            gameRepository.save(gameEntity);
-            socketWriter.write(socket, socketData);
+            gameEntity.setStarted(!gameServerService.isP2P(vers));
 
-            startGameConnection(socketWrapper, gameEntity);
+            synchronized (this) {
+                // Set userflags directly to 1 (ready) for the host
+                socketWrapper.setUserflags("1");
+            }
+
+            gameRepository.save(gameEntity);
+            socketWriter.write(socket, new SocketData("gcre", null, gameUtils.getGameInfo(gameEntity)));
+
+            startGameConnection(socketWrapper, gameEntity, true);
             personaService.who(socket, socketWrapper); // Used to set the game id
+
+            if (gameServerService.isP2P(vers)) {
+                personaService.usr(socket, socketWrapper);
+
+                // Add the game to the room
+                Room room = roomService.getRoomByVers(vers);
+                room.getGameIds().add(gameEntity.getId());
+
+                // Broadcast the game creation to people inside the room
+                socketManager.getSocketWrapperByVers(vers).stream()
+                        .filter(wrapper -> room.getPersonaIds().contains(wrapper.getPersonaEntity().getId()))
+                        .forEach(wrapper -> {
+                            socketWriter.write(wrapper.getSocket(), new SocketData("+agm", null, gameUtils.getGameInfo(gameEntity)));
+                        });
+            }
+
             try {
                 Thread.sleep(100);
             } catch (InterruptedException e) {
                 Thread.currentThread().interrupt();
             }
-            socketWriter.write(socket, new SocketData("+mgm", null, getGameInfo(gameEntity)));
+            socketWriter.write(socket, new SocketData("+mgm", null, gameUtils.getGameInfo(gameEntity)));
         }
     }
 
@@ -527,28 +691,38 @@ public class GameService {
      * @param socketWrapper The socket wrapper of current connection
      */
     public void gdel(Socket socket, SocketData socketData, SocketWrapper socketWrapper) {
-        List<GameEntity> gameEntity = gameRepository.findCurrentGameOfPersona(socketWrapper.getPersonaConnectionEntity().getId());
-        if (!gameEntity.isEmpty()) {
-            GameEntity game = gameEntity.get(0);
-            LocalDateTime now = LocalDateTime.now();
-            game.setEndTime(now);
-            gameRepository.save(game);
-            game.getGameConnections().stream().filter(connection -> null == connection.getEndTime()).forEach(report -> {
-                report.setEndTime(now);
-                gameConnectionRepository.save(report);
-            });
-        }
+        endGame(socketWrapper);
         socketWriter.write(socket, socketData);
     }
 
     /**
-     * Start session
+     * Game status update (new game, new players in the game, etc.)
+     *
+     * @param socket     The socket to write the response to
+     * @param gameEntity The game entity to start the session for
+     */
+    public void agm(Socket socket, GameEntity gameEntity) {
+        socketWriter.write(socket, new SocketData("+agm", null, gameUtils.getGameInfo(gameEntity)));
+    }
+
+    /**
+     * Join game
+     *
+     * @param socket     The socket to write the response to
+     * @param gameEntity The game entity to start the session for
+     */
+    public void mgm(Socket socket, GameEntity gameEntity) {
+        socketWriter.write(socket, new SocketData("+mgm", null, gameUtils.getGameInfo(gameEntity)));
+    }
+
+    /**
+     * Start game
      *
      * @param socket     The socket to write the response to
      * @param gameEntity The game entity to start the session for
      */
     public void ses(Socket socket, GameEntity gameEntity) {
-        socketWriter.write(socket, new SocketData("+ses", null, getGameInfo(gameEntity)));
+        socketWriter.write(socket, new SocketData("+ses", null, gameUtils.getGameInfo(gameEntity)));
     }
 
     /**
@@ -558,142 +732,14 @@ public class GameService {
      * @param socketData The socket data
      */
     public void gget(Socket socket, SocketData socketData) {
-        SocketWrapper socketWrapper = socketManager.getSocketWrapper(socket);
-        String vers = socketWrapper.getPersonaConnectionEntity().getVers();
         String ident = getValueFromSocket(socketData.getInputMessage(), "IDENT");
         Optional<GameEntity> gameEntityOpt = gameRepository.findById(Long.valueOf(ident));
         if (gameEntityOpt.isPresent()) {
             GameEntity gameEntity = gameEntityOpt.get();
-            socketWriter.write(socket, new SocketData("gget", null, getGameInfo(gameEntity, vers)));
+            socketWriter.write(socket, new SocketData("gget", null, gameUtils.getGameInfo(gameEntity)));
         } else {
             socketWriter.write(socket, new SocketData("gget", null, null));
         }
-    }
-
-    /**
-     * Get game info without version requirement
-     *
-     * @param gameEntity The game entity to get info for
-     * @return Map with game info
-     */
-    public Map<String, String> getGameInfo(GameEntity gameEntity) {
-        return getGameInfo(gameEntity, "");
-    }
-
-    /**
-     * Get game info
-     *
-     * @param gameEntity The game entity to get info for
-     * @param vers       The version of the game (used for generating OPPARAM)
-     * @return Map with game info
-     */
-    public Map<String, String> getGameInfo(GameEntity gameEntity, String vers) {
-        Long gameId = gameEntity.getId();
-        SocketWrapper hostSocketWrapperOfGame = socketManager.getHostSocketWrapperOfGame(gameId);
-
-        List<GameConnectionEntity> gameConnections = gameConnectionRepository.findByGameIdAndEndTimeIsNull(gameId);
-
-        // Workaround when there is no host (serverless patch)
-        boolean hasHost = hostSocketWrapperOfGame != null;
-        String host = hasHost ? "@" + hostSocketWrapperOfGame.getPersonaEntity().getPers() : "@brobot1";
-        int count = gameConnections.size();
-        if (!hasHost) count++;
-
-        Map<String, String> content = Stream.of(new String[][]{
-                {"IDENT", String.valueOf(Optional.ofNullable(gameEntity.getOriginalId()).orElse(gameEntity.getId()))},
-                {"NAME", gameEntity.getName()},
-                {"HOST", host},
-                // { "GPSHOST", hostSocketWrapperOfGame.getPers() },
-                {"PARAMS", gameEntity.getParams()},
-                {"PLATPARAMS", "0"},  // ???
-                {"ROOM", "1"},
-                {"CUSTFLAGS", "413082880"},
-                {"SYSFLAGS", gameEntity.getSysflags()},
-                {"COUNT", String.valueOf(count)},
-                // { "GPSREGION", "2" },
-                {"PRIV", "0"},
-                {"MINSIZE", String.valueOf(gameEntity.getMinsize())},
-                {"MAXSIZE", String.valueOf(gameEntity.getMaxsize())},
-                {"NUMPART", "1"},
-                {"SEED", "3"}, // random seed
-                {"WHEN", DateTimeFormatter.ofPattern(DATETIME_FORMAT).format(gameEntity.getStartTime())},
-                // { "GAMEPORT", String.valueOf(props.getUdpPort())},
-                // { "VOIPPORT", "9667" },
-                // { "GAMEMODE", "0" }, // ???
-                {"AUTH", gameEntity.getSysflags().equals("262656") ? "098f6bcd4621d373cade4e832627b4f6" : ""}, // Required for ranked
-
-                // loop 0x80022058 only if COUNT>=0
-
-                // another loop 0x8002225C only if NUMPART>=0
-                // { "SELF", "" },
-
-                {"SESS", "0"}, // %s-%s-%08x 0--498ea96f
-
-                {"EVID", "0"},
-                {"EVGID", "0"},
-        }).collect(Collectors.toMap(data -> data[0], data -> data[1]));
-
-        int[] idx = {0};
-
-        if (!hasHost) {
-            content.putAll(Stream.of(new String[][]{
-                    {"OPID" + idx[0], "0"},
-                    {"OPPO" + idx[0], "@brobot1"},
-                    {"ADDR" + idx[0], "127.0.0.1"},
-                    {"LADDR" + idx[0], "127.0.0.1"},
-                    {"MADDR" + idx[0], ""},
-            }).collect(Collectors.toMap(data -> data[0], data -> data[1])));
-            idx[0]++;
-        }
-
-        gameConnections.stream()
-                .sorted(Comparator.comparing(GameConnectionEntity::getId))
-                .forEach(gameConnectionEntity -> {
-                    PersonaConnectionEntity personaConnectionEntity = gameConnectionEntity.getPersonaConnection();
-                    PersonaEntity personaEntity = personaConnectionEntity.getPersona();
-                    String ipAddr = personaConnectionEntity.getAddress().replace("/", "").split(":")[0];
-                    String hostPrefix = gameConnectionEntity.isHost() ? "@" : "";
-                    content.putAll(Stream.of(new String[][]{
-                            {"OPID" + idx[0], String.valueOf(personaEntity.getId())},
-                            {"OPPO" + idx[0], hostPrefix + personaEntity.getPers()},
-                            {"ADDR" + idx[0], ipAddr},
-                            {"LADDR" + idx[0], ipAddr},
-                            {"MADDR" + idx[0], ""},
-                            {"OPPART" + idx[0], "0"},
-                            {"OPPARAM" + idx[0], generateOpParam(personaEntity, vers)},
-                            {"OPFLAG" + idx[0], "413082880"},
-                            {"OPFLAGS" + idx[0], "413082880"},
-                            {"PRES" + idx[0], "0"},
-                            {"PARTSIZE" + idx[0], String.valueOf(gameEntity.getMaxsize())},
-                            {"PARTPARAMS" + idx[0], ""},
-                    }).collect(Collectors.toMap(data -> data[0], data -> data[1])));
-                    idx[0]++;
-                });
-        return content;
-    }
-
-    /**
-     * Generate the OPPARAM for a player
-     *
-     * @param personaEntity The persona entity of the player
-     * @param vers          The version of the game
-     * @return Base64 encoded OPPARAM string
-     */
-    private String generateOpParam(PersonaEntity personaEntity, String vers) {
-        MohhPersonaStatsEntity mohhPersonaStatsEntity = mohhPersonaStatsRepository.findByPersonaIdAndVers(personaEntity.getId(), vers);
-        Long rankLong = mohhPersonaStatsRepository.getRankByPersonaIdAndVers(personaEntity.getId(), vers);
-        int rank = (rankLong != null) ? rankLong.intValue() : 0;
-        String loc = personaEntity.getAccount().getLoc();
-        String killHex = mohhPersonaStatsEntity != null ? reverseEndianness(formatIntToWord(mohhPersonaStatsEntity.getKill())) : "00000000";
-        String deathHex = mohhPersonaStatsEntity != null ? reverseEndianness(formatIntToWord(mohhPersonaStatsEntity.getDeath())) : "00000000";
-        String rankHex = reverseEndianness(formatIntToWord(rank));
-        String locHex = reverseEndianness(formatIntToWord(Integer.parseInt(stringToHex(loc.substring(loc.length() - 2)), 16)));
-        String repHex = reverseEndianness(formatIntToWord(personaEntity.getRp()));
-        String lastHex = reverseEndianness(formatIntToWord(1));
-
-        String concatenatedHex = killHex + deathHex + rankHex + locHex + repHex + lastHex;
-        byte[] byteArray = parseHexString(concatenatedHex);
-        return Base64.getEncoder().encodeToString(byteArray);
     }
 
     /**
@@ -716,14 +762,14 @@ public class GameService {
      * @param socketWrapper The socket wrapper of current connection
      * @param gameEntity    The game entity to register
      */
-    private void startGameConnection(SocketWrapper socketWrapper, GameEntity gameEntity) {
+    private void startGameConnection(SocketWrapper socketWrapper, GameEntity gameEntity, boolean isHost) {
         // Close any game report that wasn't property ended (e.g. use Dolphin save state to leave)
         endGameConnection(socketWrapper);
 
         GameConnectionEntity gameConnectionEntity = new GameConnectionEntity();
         gameConnectionEntity.setGame(gameEntity);
         gameConnectionEntity.setPersonaConnection(socketWrapper.getPersonaConnectionEntity());
-        gameConnectionEntity.setHost(socketWrapper.getIsHost().get());
+        gameConnectionEntity.setHost(isHost);
         gameConnectionEntity.setStartTime(LocalDateTime.now());
         gameConnectionRepository.save(gameConnectionEntity);
     }
@@ -737,17 +783,55 @@ public class GameService {
         if (gameConnectionEntityOpt.isPresent()) {
             GameConnectionEntity gameConnectionEntity = gameConnectionEntityOpt.get();
             GameEntity gameEntity = gameConnectionEntity.getGame();
-            if (socketWrapper.getIsHost().get()) {
-                for (GameConnectionEntity gameConnectionToClose : gameConnectionRepository.findByGameIdAndEndTimeIsNull(gameEntity.getId())) {
-                    gameConnectionToClose.setEndTime(LocalDateTime.now());
-                    gameConnectionRepository.save(gameConnectionToClose);
+
+            if (gameServerService.isP2P(gameEntity.getVers())) {
+                // If the player is the host, we need to close the game and notify all players
+                if (gameConnectionEntity.isHost()) {
+                    endGame(socketWrapper);
+                } else {
+                    gameConnectionEntity.setEndTime(LocalDateTime.now());
+                    gameConnectionRepository.save(gameConnectionEntity);
+                    // Broadcast the game leave to all connected clients in the game
+                    socketManager.getSocketWrapperByVers(gameEntity.getVers())
+                            .forEach(wrapper -> {
+                                Socket socket = wrapper.getSocket();
+                                agm(socket, gameEntity);
+                                mgm(socket, gameEntity);
+                            });
                 }
-                gameEntity.setEndTime(LocalDateTime.now());
-                gameRepository.save(gameEntity);
             } else {
-                gameConnectionEntity.setEndTime(LocalDateTime.now());
-                gameConnectionRepository.save(gameConnectionEntity);
-                updateHostInfo(gameEntity);
+                if (socketWrapper.getIsDedicatedHost().get()) {
+                    endGame(socketWrapper);
+                } else {
+                    gameConnectionEntity.setEndTime(LocalDateTime.now());
+                    gameConnectionRepository.save(gameConnectionEntity);
+                    updateHostInfo(gameEntity);
+                }
+            }
+        }
+    }
+
+    /**
+     * Ends the game and all connections to the game
+     * This is used when the game is over or when the host leaves the game
+     *
+     * @param socketWrapper The socket wrapper of current connection
+     */
+    public void endGame(SocketWrapper socketWrapper) {
+        List<GameEntity> gameEntity = gameRepository.findCurrentGameOfPersona(socketWrapper.getPersonaConnectionEntity().getId());
+        if (!gameEntity.isEmpty()) {
+            GameEntity game = gameEntity.get(0);
+            LocalDateTime now = LocalDateTime.now();
+            game.setEndTime(now);
+            gameRepository.save(game);
+            game.getGameConnections().stream().filter(connection -> null == connection.getEndTime()).forEach(report -> {
+                report.setEndTime(now);
+                gameConnectionRepository.save(report);
+            });
+
+            // For P2P games, remove the game from the room and broadcast the game deletion
+            if (gameServerService.isP2P(game.getVers())) {
+                roomService.removeGameFromRoom(game);
             }
         }
     }
