+++ src/main/resources/db/changelog/changes/006-nhl-update.sql
@@ -1,0 +1,4 @@
+ALTER TABLE core.GAME ADD COLUMN IS_STARTED boolean NOT NULL DEFAULT false;
+UPDATE core.GAME SET IS_STARTED = true WHERE IS_STARTED = false;
+
+ALTER TABLE core.GAME ADD COLUMN ROOM_ID numeric NULL;
